requirements.txt
flask
yfinance
requests
beautifulsoup4
lxml
feedparser
python-dotenv
pandas
plotly
fredapi
